Question bank page moved to desktop folder to add mobile version
